[Bridge\PhpUnit] Cleanup BC layer
